fix: [ui] fix banner bug2
fix banner bug2

Log: [ui] fix banner bug2
Bug: https://pms.uniontech.com/bug-view-256007.html
