+++ src/plugins/cooperation/core/gui/widgets/cooperationstatewidget.h
@@ -10,8 +10,6 @@
 #include "backgroundwidget.h"
 #include "global_defines.h"
 
-class QGraphicsDropShadowEffect;
-class QPropertyAnimation;
 namespace cooperation_core {
 
 class LineWidget;
@@ -94,44 +92,6 @@ private:
     QLabel *ipLabel { nullptr };
     QTimer *timer { nullptr };
 };
-
-class FirstTipWidget : public QWidget
-{
-    Q_OBJECT
-public:
-    explicit FirstTipWidget(QWidget *parent = nullptr);
-
-    void setVisible(bool visible) override;
-    void themeTypeChanged();
-
-protected:
-    void showEvent(QShowEvent *event) override;
-    void hideEvent(QHideEvent *event) override;
-
-private:
-    void initUI();
-    void initbackgroundFrame();
-    void inittipBtn();
-
-    QFrame *backgroundFrame { nullptr };
-    QLabel *bannerLabel { nullptr };
-    QLabel *action { nullptr };
-    QGraphicsDropShadowEffect *shadowEffect { nullptr };
-    QList<QLabel *> lineBalls { nullptr };
-    LineWidget *line { nullptr };
-    CooperationIconButton *tipBtn { nullptr };
-};
-
-class LineWidget : public QWidget
-{
-public:
-    LineWidget(QWidget *parent = nullptr)
-        : QWidget(parent) {}
-
-protected:
-    void paintEvent(QPaintEvent *event) override;
-};
-
 }   // namespace cooperation_core
 
 #endif   // COOPERATIONSTATEWIDGET_H
